fix drivers, addressees, userbuilder and others

+++ src/Lab3/Displays/Color.cs
@@ -4,12 +4,12 @@ public record Color
 {
     public Color(int red, int green, int blue)
     {
-        Red = red;
-        Green = green;
-        Blue = blue;
+        Red = (byte)red;
+        Green = (byte)green;
+        Blue = (byte)blue;
     }
 
-    public int Red { get; }
-    public int Green { get; }
-    public int Blue { get; }
+    public byte Red { get; }
+    public byte Green { get; }
+    public byte Blue { get; }
 }
